added check for if certificate is already downloaded

+++ setup.sh
@@ -7,7 +7,19 @@ if [ ! -d "$folder" ]; then
   mkdir -p $folder
 fi
 
-curl 'https://itswiki.compute.dtu.dk/images/0/07/Eduroam_aug2020.pem' -o $folder/dtusecure.pem
+cert="$folder/dtusecure.pem"
+
+if [ -f "$cert" ]; then
+    read -p "Certifacte already exists. Do you wish to redownload? [y/N] " answer
+
+    if [[ $answer == "y" || $answer == "Y" ]]; then
+        echo "Downloading certificate: "
+        curl 'https://itswiki.compute.dtu.dk/images/0/07/Eduroam_aug2020.pem' -o $folder/dtusecure.pem
+    else
+        echo "Skipping download of certificate..."
+    fi
+fi
+
 
 read -r -p "Username: " username
 read -r -p "Password: " -s password
@@ -21,4 +33,4 @@ nmcli connection add \
  wifi-sec.key-mgmt wpa-eap 802-1x.eap peap 802-1x.phase2-auth mschapv2 \
  802-1x.identity $username 802-1x.password $password \
  802-1x.anonymous-identity "anonymous@dtu.dk" \
- 802-1x.ca-cert $folder/dtusecure.pem
+ 802-1x.ca-cert $cert
